Removed return value from IE encoding
The interface{} return value in EncodeToIEDataType() is identical to the 
input parameter val in most scenarios. Returning this as a new interface 
pointer results in an unnecesary heap allocation for each IE in each 
record. In some cases, such as IP addresses or floats, the return value 
is slighly different than the input (IPs are cast to ipv4 or ipv6, 
floats are cast to their binary representation), but this modified value 
is not used except in test cases.
Results in ~5%-10% improvement in runtime and removes 1 allocation per 
IE in the record